+++ runtime/common/src/lib.rs
@@ -6,8 +6,6 @@ pub use parachains_common::{AccountId, AuraId, Balance, Block, BlockNumber, Hash
 pub use polkadot_primitives::MAX_POV_SIZE;
 use sp_runtime::Perbill;
 
-pub mod proxy;
-
 /// Nonce for an account
 pub type Nonce = u32;
 
@@ -75,3 +73,75 @@ mod async_backing_params {
 	pub const RELAY_CHAIN_SLOT_DURATION_MILLIS: u32 = 6000;
 }
 pub use async_backing_params::*;
+
+/// Proxy commons for Pop runtimes
+pub mod proxy {
+
+	use codec::{Decode, Encode, MaxEncodedLen};
+	use frame_support::parameter_types;
+	use sp_runtime::RuntimeDebug;
+
+	use super::{deposit, Balance};
+
+	parameter_types! {
+		// One storage item; key size 32, value size 8; .
+		pub const ProxyDepositBase: Balance = deposit(1, 40);
+		// Additional storage item size of 33 bytes.
+		pub const ProxyDepositFactor: Balance = deposit(0, 33);
+		pub const MaxProxies: u16 = 32;
+		// One storage item; key size 32, value size 16
+		pub const AnnouncementDepositBase: Balance = deposit(1, 48);
+		pub const AnnouncementDepositFactor: Balance = deposit(0, 66);
+		pub const MaxPending: u16 = 32;
+	}
+
+	/// The type used to represent the kinds of proxying allowed.
+	#[derive(
+		Copy,
+		Clone,
+		Eq,
+		PartialEq,
+		Ord,
+		PartialOrd,
+		Encode,
+		Decode,
+		RuntimeDebug,
+		MaxEncodedLen,
+		scale_info::TypeInfo,
+	)]
+	pub enum ProxyType {
+		/// Fully permissioned proxy. Can execute any call on behalf of _proxied_.
+		Any,
+		/// Can execute any call that does not transfer funds or assets.
+		NonTransfer,
+		/// Proxy with the ability to reject time-delay proxy announcements.
+		CancelProxy,
+		/// Assets proxy. Can execute any call from `assets`, **including asset transfers**.
+		Assets,
+		/// Owner proxy. Can execute calls related to asset ownership.
+		AssetOwner,
+		/// Asset manager. Can execute calls related to asset management.
+		AssetManager,
+		/// Collator selection proxy. Can execute calls related to collator selection mechanism.
+		Collator,
+	}
+	impl Default for ProxyType {
+		fn default() -> Self {
+			Self::Any
+		}
+	}
+
+	impl ProxyType {
+		pub fn is_superset(s: &ProxyType, o: &ProxyType) -> bool {
+			match (s, o) {
+				(x, y) if x == y => true,
+				(ProxyType::Any, _) => true,
+				(_, ProxyType::Any) => false,
+				(ProxyType::Assets, ProxyType::AssetOwner) => true,
+				(ProxyType::Assets, ProxyType::AssetManager) => true,
+				(ProxyType::NonTransfer, ProxyType::Collator) => true,
+				_ => false,
+			}
+		}
+	}
+}
